+++ src/js/blueprint.js
@@ -2,31 +2,34 @@ class Blueprint {
     constructor(app, name) {
         this.app = app;
         this.name = name;
+        this.methods = {};
         this.nodes = [];
         this.links = [];
 
         setTimeout(() => {
+            this.createMethod("Constructor");
             this.createNode({ x: -400, y: 0 }, 0);
-            this.createNode({ x: 0, y: -50 }, 1);
+            this.createNode({ x: 0, y: -50 }, 10);
             this.createNode({ x: -400, y: 120 }, 100);
-            this.createNode({ x: 300, y: 120 }, 1);
+            this.createNode({ x: 300, y: 0 }, 1);
         }, 0);
     }
 
     open() {
         this.refreshNodes();
+        this.refreshLinks();
     }
 
     close() {
         this.nodes.forEach((node) => node.hide());
     }
 
-    refreshNodes() {
-        this.nodes.filter((node) => !node.displayed).forEach((node) => node.show());
+    createMethod(name) {
+        
     }
 
-    refreshLinks() {
-        this.links.filter((link) => !link.displayed).forEach((link) => link.show());
+    removeMethod(uid) {
+
     }
 
     createNode(pos, id) {
@@ -40,12 +43,19 @@ class Blueprint {
     removeNode(uid) {
         this.nodes = this.nodes.filter((node) => {
             if (node.uid == uid) {
-                node.hide();
+                if (node.id <= 1) {
+                    this.app.error(`You can't remove a "${node.data.title}" node`);
+                    return true;
+                } else node.hide();
                 return false;
             } else return true;
         });
         this.refreshNodes();
         this.app.getLinksByNodeUID(uid).forEach((link) => this.removeLink(link.uid));
+    }
+
+    refreshNodes() {
+        this.nodes.filter((node) => !node.displayed).forEach((node) => node.show());
     }
 
     linkNodes(returnNode, returnIndex, parameterNode, parameterIndex) {
@@ -72,15 +82,18 @@ class Blueprint {
     }
 
     removeLink(uid) {
+        const toHide = [];
         this.links = this.links.filter((link) => {
             if (link.uid == uid) {
-                // if (link.data.returnNode.uid)
-                // console.log(link.data);
-                
-                link.hide();
+                toHide.push(link);
                 return false;
             } else return true;
         });
+        toHide.forEach((link) => link.hide());
         this.refreshLinks();
+    }
+
+    refreshLinks() {
+        this.links.filter((link) => !link.displayed).forEach((link) => link.show());
     }
 };