+++ src/js/contextmenu.js
@@ -18,9 +18,11 @@ class ContextMenu {
         input.addEventListener("input", (e) => {
             nodeList.innerHTML = "";
             Object.entries(this.app.nodes).forEach(([id, node]) => {
+                if (id <= 1) return;
                 if (!node.title.toLowerCase().includes(e.target.value.toLowerCase())) return;
                 const li = document.createElement("li");
                 li.textContent = node.title;
+                li.setAttribute("node-id", id);
                 nodeList.appendChild(li);
 
                 li.addEventListener("click", () => {
@@ -28,6 +30,12 @@ class ContextMenu {
                     this.app.destroyContextMenu();
                 });
             });
+        });
+        input.addEventListener("keydown", (e) => {
+            if (e.key != "Enter") return;
+            if (nodeList.children.length != 1) return;
+            this.app.currentBlueprint.createNode(this.app.getCursorPos({ x: x, y: y }), [...nodeList.children][0].getAttribute("node-id"));
+            this.app.destroyContextMenu();
         });
 
         input.dispatchEvent(new Event("input"));