+++ src/js/link.js
@@ -6,6 +6,8 @@ class Link {
         this.displayed = false;
         this.uid = this.app.getNextLinkUID(blueprint);
         this.element = null;
+        this.returnLink = null;
+        this.parameterLink = null;
     }
 
     show() {
@@ -15,17 +17,17 @@ class Link {
         const returnNode = this.blueprint.nodes.filter((node) => node.uid == this.data.returnNode.uid)[0];
         const parameterNode = this.blueprint.nodes.filter((node) => node.uid == this.data.parameterNode.uid)[0];
 
-        const returnLink = returnNode.element.querySelector("ul.output-container").children[this.data.returnNode.index].querySelector("div.linker");
-        const parameterLink = parameterNode.element.querySelector("ul.input-container").children[this.data.parameterNode.index].querySelector("div.linker");
+        this.returnLink = returnNode.element.querySelector("ul.output-container").children[this.data.returnNode.index].querySelector("div.linker");
+        this.parameterLink = parameterNode.element.querySelector("ul.input-container").children[this.data.parameterNode.index].querySelector("div.linker");
 
-        const retRect = returnLink.getBoundingClientRect();
-        const parRect = parameterLink.getBoundingClientRect();
+        const retRect = this.returnLink.getBoundingClientRect();
+        const parRect = this.parameterLink.getBoundingClientRect();
 
-        const returnX = returnNode.pos.x + returnLink.offsetLeft + retRect.width / 2;
-        const returnY = returnNode.pos.y + returnLink.offsetTop + retRect.height / 2;
+        const returnX = returnNode.pos.x + this.returnLink.offsetLeft + retRect.width / 2;
+        const returnY = returnNode.pos.y + this.returnLink.offsetTop + retRect.height / 2;
 
-        const parameterX = parameterNode.pos.x + parameterLink.offsetLeft + parRect.width / 2;
-        const parameterY = parameterNode.pos.y + parameterLink.offsetTop + parRect.height / 2;
+        const parameterX = parameterNode.pos.x + this.parameterLink.offsetLeft + parRect.width / 2;
+        const parameterY = parameterNode.pos.y + this.parameterLink.offsetTop + parRect.height / 2;
 
         const xDiff = Math.abs(returnX - parameterX);
         const yDiff = Math.abs(returnY - parameterY);
@@ -38,8 +40,8 @@ class Link {
         const width = Math.sqrt(Math.pow(xDiff, 2) + Math.pow(yDiff, 2));
         const angle = Math.atan2(yDiff, xDiff) * (180 / Math.PI) * ((returnY > parameterY) ? -1 : 1) * ((returnX > parameterX) ? -1 : 1);
 
-        returnLink.classList.add("linked");
-        parameterLink.classList.add("linked");
+        this.returnLink.classList.add("linked");
+        this.parameterLink.classList.add("linked");
         
         this.element = document.createElement("div");
         this.element.classList.add("link");
@@ -54,6 +56,8 @@ class Link {
     hide() {
         if (!this.app.isCurrentBP(this.blueprint)) return;
         if (this.element) this.element.remove();
+        if (this.returnLink) if (this.app.getLinkerNbLinks(this.returnLink) == 0) this.returnLink.classList.remove("linked");
+        if (this.parameterLink) if (this.app.getLinkerNbLinks(this.parameterLink) == 0) this.parameterLink.classList.remove("linked");
         this.displayed = false;
     }
 };